requirements.txt
python-telegram-bot==20.*
python-dotenv
torch
torchvision
opencv-python
Pillow
